Move PluginConfigPage to api.preferences
