+++ spyder/plugins/ipythonconsole.py
@@ -47,7 +47,7 @@ from spyder.config.base import (_, DEV, get_home_dir, get_module_path,
                                 get_module_source_path)
 from spyder.config.main import CONF
 from spyder.api.plugins import SpyderPluginWidget
-from spyder.plugins.configdialog import PluginConfigPage
+from spyder.api.preferences import PluginConfigPage
 from spyder.py3compat import (iteritems, PY2, to_binary_string,
                               to_text_string)
 from spyder.utils.qthelpers import create_action
