+++ spyder/plugins/variableexplorer.py
@@ -14,7 +14,7 @@ from qtpy.QtWidgets import QGroupBox, QStackedWidget, QVBoxLayout, QWidget
 from spyder.config.base import _
 from spyder.config.main import CONF
 from spyder.api.plugins import SpyderPluginMixin
-from spyder.plugins.configdialog import PluginConfigPage
+from spyder.api.preferences import PluginConfigPage
 from spyder.utils import programs
 from spyder.utils import icon_manager as ima
 from spyder.widgets.variableexplorer.namespacebrowser import NamespaceBrowser
